--- jsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "src"
-  },
-  "exclude": ["node_modules", "build"],
-  "include": ["src"]
-}
