+++ package.json
@@ -6,14 +6,19 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "@types/jest": "^28.1.1",
+    "@types/node": "^17.0.41",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
     "bootstrap": "^5.1.3",
     "eslint": "^7.11.0",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-bootstrap": "^2.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "4.0.3",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.7.3",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
