+++ src/components/Cards/MainCard/MainCard.tsx
@@ -3,8 +3,8 @@ import './MainCard.css'
 import Card from 'react-bootstrap/Card'
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import LogoLeft from 'assets/svg/logo-left.svg'
-import LogoRight from 'assets/svg/logo-right.svg'
+import LogoLeft from '../../../assets/svg/logo-left.svg'
+import LogoRight from '../../../assets/svg/logo-right.svg'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
