+++ src/components/Cards/MembersCard/MembersCard.tsx
@@ -3,7 +3,7 @@ import './MembersCard.css'
 import Card from 'react-bootstrap/Card'
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import Image from 'assets/img/placeholder.jpeg'
+import Image from '../../../assets/img/placeholder.jpeg'
 import { Link } from 'react-router-dom'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
