+++ src/frontend-page/team-page/Frontend.tsx
@@ -1,14 +1,14 @@
-import 'frontend-page/TeamPage.css'
+import '../TeamPage.css'
 
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import MainCard from 'components/Cards/MainCard/MainCard'
-import Image from 'assets/img/placeholder.jpeg'
-import MembersCard from 'components/Cards/MembersCard/MembersCard'
+import MainCard from '../../components/Cards/MainCard/MainCard'
+import Image from '../../assets/img/placeholder.jpeg'
+import MembersCard from '../../components/Cards/MembersCard/MembersCard'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
 
-export const FrontendTeamPage = () => {
+export const FrontendTeamPage: any = () => {
   const card = {
     team: 'FrontendTeamPage',
   }
