+++ src/frontend-page/team-page/Outreach.tsx
@@ -1,20 +1,20 @@
-import 'frontend-page/TeamPage.css'
+import '../TeamPage.css'
 
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import MainCard from 'components/Cards/MainCard/MainCard'
-import Image from 'assets/img/placeholder.jpeg'
-import MembersCard from 'components/Cards/MembersCard/MembersCard'
+import MainCard from '../../components/Cards/MainCard/MainCard'
+import Image from '../../assets/img/placeholder.jpeg'
+import MembersCard from '../../components/Cards/MembersCard/MembersCard'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
 
-export const ProjectTeamPage = () => {
+export const OutreachTeamPage = () => {
   const card = {
-    team: 'ProjectTeamPage',
+    team: 'OutreachTeamPage',
   }
 
   const members = {
-    teamName: 'Project Team',
+    teamName: 'Outreach Team',
     leader: { name: 'FirstName LastName' },
     member: [
       { name: 'FirstName LastName' },
@@ -63,4 +63,4 @@ export const ProjectTeamPage = () => {
   )
 }
 
-export default ProjectTeamPage
+export default OutreachTeamPage
