+++ src/types/react-app-env.d.ts
@@ -1,0 +1,4 @@
+/// <reference types='react-scripts'/>
+
+declare module '*.module.css';
+declare module '*..module.scss'
